refactored for go optimization
