+++ handlers.go
@@ -1,0 +1,51 @@
+package main
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+)
+
+// ImageUploadHandler handles the image upload requests.
+func ImageUploadHandler(c *gin.Context) {
+	file, err := c.FormFile("image")
+	if err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	filePath := "path/to/save/" + file.Filename
+	if err := c.SaveUploadedFile(file, filePath); err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{"message": "File Uploaded successfully."})
+}
+
+// CreateNFTHandler handles the NFT creation requests.
+func CreateNFTHandler(c *gin.Context) {
+	type RequestBody struct {
+		Text string `json:"text"`
+	}
+
+	var requestBody RequestBody
+	if err := c.BindJSON(&requestBody); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	// Add logic here for NFT creation using the requestBody.Text
+
+	c.JSON(http.StatusOK, gin.H{"message": "NFT creation initiated."})
+}
+
+// TrainGANHandler handles requests to train the GAN.
+func TrainGANHandler(c *gin.Context) {
+	err := TrainGAN() // This function should be defined in your GAN integration logic
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to train GAN"})
+		return
+	}
+	c.JSON(http.StatusOK, gin.H{"message": "GAN training initiated"})
+}
